change i and j type to `int`

+++ modules/line_descriptor/src/binary_descriptor.cpp
@@ -623,9 +623,9 @@ void BinaryDescriptor::computeImpl( const Mat& imageSrc, std::vector<KeyLine>& k
   }
 
   /* delete useless OctaveSingleLines */
-  for ( size_t i = 0; i < sl.size(); i++ )
+  for ( int i = 0; i < sl.size(); i++ )
   {
-    for ( size_t j = sl[i].size() - 1; j >= 0; --j )
+    for ( int j = sl[i].size() - 1; j >= 0; --j )
     {
       //if( (int) ( sl[i][j] ).octaveCount > params.numOfOctave_ )
       if( (int) ( sl[i][j] ).octaveCount > octaveIndex )
